Allow building anywhere.
The EPICS build process is a bit tricky. EPICS base doesn't pose any problem, but the support modules require hardcoding the values of location of base (EPICS_BASE) and the other support modules (SUPPORT) in a RELEASE file. It is not possible to pass these values as variables, in fact the compilation scripts check their source and aborts if they don't come from a file. As a work-around, our Makefile creates the file RELEASE.local with the values provided by the user, to make EPICS happy. This file is included in the final target directory. The result of the compilation still includes hardcoded values, so it is not easy to move it to another location. If such a need arises, it is better to recompile again.